fix(lifecycle): platform check for default install scripts
This is a continuation of #45, a side-effect of the provided fix for it
in #46 and was introduced in `cipm@1.6.2`.

For the case where a default `install` script is used and a
`binding.gyp` file is present in  the package root, `npm ci` will fail
with packages that target a different platform that the one currently
running.

Fixes #49